Distribute stuff from `FoundationExtensions` to separate files
