+++ ACKategories/NSAttributedStringExtensions.swift
@@ -1,11 +1,3 @@
-//
-//  NSAttributedStringExtensions.swift
-//  ACKategories
-//
-//  Created by Josef Dolezal on 04/02/2018.
-//  Copyright © 2018 Josef Dolezal. All rights reserved.
-//
-
 import Foundation
 
 public extension NSAttributedString {
