+++ ACKategories/NSMutableParagraphStyleExtensions.swift
@@ -1,11 +1,3 @@
-//
-//  NSMutableParagraphStyleExtensions.swift
-//  ACKategories
-//
-//  Created by Jakub Olejník on 13/04/2018.
-//  Copyright © 2018 Josef Dolezal. All rights reserved.
-//
-
 import UIKit
 
 extension NSMutableParagraphStyle {
