+++ ACKategories/UIButtonExtensions.swift
@@ -1,11 +1,3 @@
-//
-//  UIButtonExtensions.swift
-//  ACKategories
-//
-//  Created by Jakub Olejník on 06/02/2018.
-//  Copyright © 2018 Josef Dolezal. All rights reserved.
-//
-
 import UIKit
 
 extension UIButton {
@@ -16,5 +8,14 @@ extension UIButton {
         
         imageEdgeInsets = UIEdgeInsets(top: -(totalHeight - imageSize.height), left: 0, bottom: 0, right: -titleSize.width)
         titleEdgeInsets = UIEdgeInsets(top: 0, left: -imageSize.width, bottom: -(totalHeight - titleSize.height), right: 0)
+    }
+}
+
+extension UIButton {
+    open override var intrinsicContentSize: CGSize {
+        let size = super.intrinsicContentSize
+        let width = size.width + titleEdgeInsets.left + titleEdgeInsets.right
+        let height = size.height + titleEdgeInsets.top + titleEdgeInsets.bottom
+        return CGSize(width: width, height: height)
     }
 }
